requirements.txt
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0
flask>=3.0.0
python-dotenv>=1.0.0
